Fix the macOS and MSVC builds
Two portability bugs from the transport hardening commits, both of
which Linux is structurally incapable of catching.

MSVC: net.c used uint64_t in the platform shim -- the monotonic clock
-- before anything had declared it. The POSIX socket headers happen to
drag the fixed-width types in, so on Linux and macOS the file compiled;
winsock2.h has no such side effect, so MSVC failed with a bare syntax
error at the function signature. stdint.h (and stddef.h, for the same
reason) now come first, above the shim, with a note saying why.

macOS: struct timeval's tv_usec is `int` on Darwin and `long` on Linux
and Windows, so `(long)(ms % 1000u) * 1000L` was exact on two platforms
and a 64-to-32 narrowing on the third. Assigning from `int` narrows
nowhere -- the value is under a million by construction.

Also in the same function, select()'s nfds is now skipped entirely on
Winsock, which ignores it: computing `s + 1` for a pointer-width SOCKET
handle was inventing a narrowing conversion to satisfy an argument
nobody reads.

And the two %zu conversions introduced with the snapshot logging are
now %llu with an explicit cast. %zu appeared nowhere else in this
codebase; the established convention is the cast, and following it is
worth more here than being terse on a specifier whose CRT support this
project has never exercised.

Checked by building the whole tree with clang as well as gcc, and by
compiling the touched files under -Wshorten-64-to-32 and
-Wsign-conversion on top of the project's flags. Neither reproduces a
Darwin type difference, which is the real lesson: this class needs CI,
not a stricter local build.

+++ src/game.c
@@ -526,10 +526,10 @@ static int save_write(const GameState *gs, const char *path,
     }
 
     if (snap)
-        sim_log("Checkpoint written to %s (tick %llu, %zu-byte snapshot, "
+        sim_log("Checkpoint written to %s (tick %llu, %llu-byte snapshot, "
                 "%d pending commands, %d intents)",
-                path, (unsigned long long)gs->sim_tick_no, snap_len, n,
-                gs->intent_count);
+                path, (unsigned long long)gs->sim_tick_no,
+                (unsigned long long)snap_len, n, gs->intent_count);
     else
         sim_log("Game saved to %s (seed %u, tick %llu, %d commands, %d intents)",
                 path, gs->world_seed,
